requirements.txt
-i https://pypi.org/simple
absl-py==2.1.0; python_version >= '3.7'
certifi==2024.6.2; python_version >= '3.6'
charset-normalizer==3.3.2; python_full_version >= '3.7.0'
cloudpickle==3.0.0; python_version >= '3.8'
filelock==3.15.1; python_version >= '3.8'
fsspec==2024.6.0; python_version >= '3.8'
grpcio==1.64.1; python_version >= '3.8'
huggingface-hub==0.23.4; python_full_version >= '3.8.0'
idna==3.7; python_version >= '3.5'
jinja2==3.1.4; python_version >= '3.7'
markdown==3.6; python_version >= '3.8'
markupsafe==2.1.5; python_version >= '3.7'
mpmath==1.3.0
networkx==3.3; python_version >= '3.10'
numpy==1.26.4; python_version >= '3.9'
nvidia-cublas-cu12==12.1.3.1; platform_system == 'Linux' and platform_machine == 'x86_64'
nvidia-cuda-cupti-cu12==12.1.105; platform_system == 'Linux' and platform_machine == 'x86_64'
nvidia-cuda-nvrtc-cu12==12.1.105; platform_system == 'Linux' and platform_machine == 'x86_64'
nvidia-cuda-runtime-cu12==12.1.105; platform_system == 'Linux' and platform_machine == 'x86_64'
nvidia-cudnn-cu12==8.9.2.26; platform_system == 'Linux' and platform_machine == 'x86_64'
nvidia-cufft-cu12==11.0.2.54; platform_system == 'Linux' and platform_machine == 'x86_64'
nvidia-curand-cu12==10.3.2.106; platform_system == 'Linux' and platform_machine == 'x86_64'
nvidia-cusolver-cu12==11.4.5.107; platform_system == 'Linux' and platform_machine == 'x86_64'
nvidia-cusparse-cu12==12.1.0.106; platform_system == 'Linux' and platform_machine == 'x86_64'
nvidia-nccl-cu12==2.20.5; platform_system == 'Linux' and platform_machine == 'x86_64'
nvidia-nvjitlink-cu12==12.5.40; python_version >= '3'
nvidia-nvtx-cu12==12.1.105; platform_system == 'Linux' and platform_machine == 'x86_64'
opencv-python==4.10.0.82; python_version >= '3.6'
packaging==24.1; python_version >= '3.8'
pillow==10.3.0; python_version >= '3.8'
protobuf==4.25.3; python_version >= '3.8'
pyyaml==6.0.1; python_version >= '3.6'
requests==2.32.3; python_version >= '3.8'
safetensors==0.4.3; python_version >= '3.7'
setuptools==70.0.0; python_version >= '3.8'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
submitit==1.5.1; python_version >= '3.8'
sympy==1.12.1; python_version >= '3.8'
tensorboard==2.17.0; python_version >= '3.9'
tensorboard-data-server==0.7.2; python_version >= '3.7'
timm==1.0.3; python_version >= '3.8'
torch==2.3.0; python_full_version >= '3.8.0'
torchvision==0.18.0; python_version >= '3.8'
tqdm==4.66.4; python_version >= '3.7'
triton==2.3.0; python_version < '3.12' and platform_system == 'Linux' and platform_machine == 'x86_64'
typing-extensions==4.12.2; python_version >= '3.8'
urllib3==2.2.1; python_version >= '3.8'
werkzeug==3.0.3; python_version >= '3.8'
